pyproject.toml
[tool.poetry]
name = "keyring-pass"
version = "0.9.2"
description = "https://www.passwordstore.org/ backend for https://pypi.org/project/keyring/"
authors = ["Krzysztof Nazarewski <3494992+nazarewk@users.noreply.github.com>"]
homepage = "https://github.com/nazarewk/keyring_pass"
repository = "https://github.com/nazarewk/keyring_pass"
license = "MIT"
readme = "README.md"
keywords = [
    "keyring",
    "pass",
]
classifiers = []
packages = [{ include = "keyring_pass" }]

[tool.poetry.dependencies]
python = ">=3.7" # required by `jaraco.classes`
keyring = ">=23.9.3"
jaraco-classes = ">=3.2.3"

[tool.poetry.plugins."keyring.backends"]
pass = "keyring_pass"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
